refactor(api): move codeSample to lib and use it for every documented route
It was generic but sat in routers/admin/shared.ts, while six routes in index.ts,
v1.ts and waitlist.ts still hand-wrote the x-codeSamples block and its cast, and
the api-endpoint skill told every new route to import it from the admin resource
directory. It now lives in lib/openapi.ts beside error.ts and paging.ts, all 15
documented routes use it, and no hand-written block is left.

Also corrects the mount-order comment in admin/index.ts: each feature gate sits
inside its own router, so the order there carries no gating.
