+++ .agents/skills/api-endpoint/SKILL.md
@@ -56,7 +56,7 @@ export const exampleRouter = new Hono().post(
 ```
 
 - Spread the matching error-response set into `responses` so its shape shows in the Scalar docs: `...validationErrorResponses` (400) for a validated route, `...authErrorResponses` (401) for an auth route, `...forbiddenErrorResponses` (403) for an admin route, `...notFoundErrorResponses` (404) for one addressing a row by id, and `...conflictErrorResponses` (409) for one that can lose a race to a unique constraint or to a concurrent edit. 429 and 500 are added globally in `index.ts`, so never per route.
-- Spread `codeSample(...)` from `routers/admin/shared.ts` so Scalar shows the `hono/client` usage (the template above omits it). It writes the `x-codeSamples` block and its cast, so a route supplies only the snippet.
+- Spread `codeSample(...)` from `lib/openapi.ts` so Scalar shows the `hono/client` usage (the template above omits it). It writes the `x-codeSamples` block and its cast, so a route supplies only the snippet.
 - Auth-protected routes go in `v1.ts`, behind `authMiddleware` from `@/middlewares` with `Variables: Session` so `c.get("session")`/`c.get("user")` are typed. A public route gets its own router.
 - Console routes go in `routers/admin/`, one file per resource re-exported through `admin/index.ts` and mounted at `/admin` inside `v1.ts` behind the console gate (stacked after `authMiddleware`). Every route there is a console surface for admins, whether that is who reaches the console, the rules that let them, the trail of those changes, or the waitlist, so the router mounts `consoleAdminMiddleware` and requires admin throughout. That middleware is one instance of `requireConsoleRole(minimum)`, the factory to call if a surface ever wants a lower rung. It re-reads the session with `disableCookieCache: true` and also refuses a banned user, so a demotion or ban lands on the next request; never gate on the cached session's role.
 - A route the browser calls with a method other than GET or POST needs that method in the `cors()` `allowMethods` list in `index.ts`, and in the OpenAPI `defaultOptions` beside it, or the preflight fails and the request never leaves the page while its 429 and 500 go undocumented. curl will not catch the preflight.
